+++ toys-core/CHANGELOG.md
@@ -2,12 +2,16 @@
 
 ### 0.3.4 / TBD
 
-* IMPROVED: gem_build template can suppress interactive confirmation
+* CHANGED: Removed Loader#execute, and returned remaining args from Loader#lookup.
+* CHANGED: Wrapped most errors with Toys::ContextualError
+* IMPROVED: gem_build template can suppress interactive confirmation.
+* IMPROVED: Logger colors the header when possible.
+* ADDED: CLI now takes an error handler to report most errors.
 
 ### 0.3.3 / 2018-05-09
 
 * CHANGED: Renamed file_utils helper to fileutils.
-* CHANGED: Renamed doc: parameter to docs:
+* CHANGED: Renamed `doc:` parameter to `docs:`.
 * CHANGED: SwitchDefinition has separate fields for acceptor and docs.
 * CHANGED: Description and long description are now arrays of strings.
 * FIXED: Documentation strings that begin with "--" no longer cause problems.