+++ toys/CHANGELOG.md
@@ -2,8 +2,10 @@
 
 ### 0.3.4 / TBD
 
-* IMPROVED: gem_build template can suppress interactive confirmation
-* IMPROVED: system update builtin can optionally ask for confirmation
+* IMPROVED: gem_build template can suppress interactive confirmation.
+* IMPROVED: system update builtin can optionally ask for confirmation.
+* IMPROVED: Error reporting is significantly improved.
+* IMPROVED: Logger colors the header when possible.
 
 ### 0.3.3 / 2018-05-09